+++ packages/web3-eth-abi/src/api/functions_api.ts
@@ -17,8 +17,8 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 
 import { AbiError } from 'web3-errors';
 import { sha3Raw } from 'web3-utils';
+import { AbiFunctionFragment } from 'web3-types';
 import { isAbiFunctionFragment, jsonInterfaceMethodToString } from '../utils';
-import { AbiFunctionFragment } from '../types';
 import { encodeParameters } from './parameters_api';
 
 /**
