+++ packages/web3-eth-abi/src/index.ts
@@ -15,7 +15,6 @@ You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-export * from './types';
 export * from './api/errors_api';
 export * from './api/events_api';
 export * from './api/functions_api';
