+++ packages/web3-eth-contract/src/contract.ts
@@ -26,6 +26,14 @@ import {
 	SendTransactionEvents,
 } from 'web3-eth';
 import {
+	encodeEventSignature,
+	encodeFunctionSignature,
+	isAbiErrorFragment,
+	isAbiEventFragment,
+	isAbiFunctionFragment,
+	jsonInterfaceMethodToString,
+} from 'web3-eth-abi';
+import {
 	AbiConstructorFragment,
 	AbiErrorFragment,
 	AbiEventFragment,
@@ -38,21 +46,13 @@ import {
 	ContractMethod,
 	ContractMethodInputParameters,
 	ContractMethodOutputParameters,
-	encodeEventSignature,
-	encodeFunctionSignature,
-	FilterAbis,
-	isAbiErrorFragment,
-	isAbiEventFragment,
-	isAbiFunctionFragment,
-	jsonInterfaceMethodToString,
-} from 'web3-eth-abi';
-import {
 	Address,
 	BlockNumberOrTag,
 	BlockTags,
 	Bytes,
 	EthExecutionAPI,
 	Filter,
+	FilterAbis,
 	HexString,
 	LogsInput,
 	Mutable,
