+++ packages/web3-eth-contract/src/encoding.ts
@@ -17,13 +17,20 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 
 import { DataFormat, DEFAULT_RETURN_FORMAT, format, isNullish } from 'web3-utils';
 
-import { LogsInput, BlockNumberOrTag, Filter, HexString, Topic, Numbers } from 'web3-types';
-
 import {
 	AbiConstructorFragment,
 	AbiErrorFragment,
 	AbiEventFragment,
 	AbiFunctionFragment,
+	LogsInput,
+	BlockNumberOrTag,
+	Filter,
+	HexString,
+	Topic,
+	Numbers,
+} from 'web3-types';
+
+import {
 	decodeLog,
 	decodeParameters,
 	encodeErrorSignature,
