+++ packages/web3/src/web3.ts
@@ -17,7 +17,6 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 // eslint-disable-next-line max-classes-per-file
 import { Web3Context } from 'web3-core';
 import Web3Eth from 'web3-eth';
-import { ContractAbi } from 'web3-eth-abi';
 import Contract, { ContractInitOptions } from 'web3-eth-contract';
 import { ENS, registryAddresses } from 'web3-eth-ens';
 import Iban from 'web3-eth-iban';
@@ -25,7 +24,7 @@ import Personal from 'web3-eth-personal';
 import Net from 'web3-net';
 import * as utils from 'web3-utils';
 import { isNullish } from 'web3-utils';
-import { EthExecutionAPI, Address, SupportedProviders } from 'web3-types';
+import { Address, ContractAbi, EthExecutionAPI, SupportedProviders } from 'web3-types';
 import abi from './abi';
 import { initAccountsForContext } from './accounts';
 import { Web3EthInterface } from './types';
