+++ tools/web3-plugin-example/src/contract_method_wrappers.ts
@@ -15,9 +15,8 @@ You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 import { Web3PluginBase } from 'web3-core';
-import { ContractAbi } from 'web3-eth-abi';
 import Contract from 'web3-eth-contract';
-import { Address, Numbers } from 'web3-types';
+import { Address, ContractAbi, Numbers } from 'web3-types';
 import { DataFormat, DEFAULT_RETURN_FORMAT, format, numberToHex } from 'web3-utils';
 
 import { ERC20TokenAbi } from './ERC20Token';
